next: wrap searchparams in suspense block
